backend/requirements.txt
# API & Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# HTTP & Crawling
httpx>=0.27.0

# HTML Parsing
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
lxml>=5.0.0

# Search & Retrieval
rank-bm25>=0.2.2
nltk>=3.8.0

# Spell correction
pyspellchecker>=0.8.0

# Embeddings & Reranking
sentence-transformers>=2.2.0

# Vector Store (optional - falls back to memory stub if not installed)
chromadb>=0.4.0

# Config & Utils
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Optional: LLM query expansion (set OPENAI_API_KEY)
openai>=1.0.0
